Renames components
