+++ client/app/queue/JudgeReviewTaskListView.jsx
@@ -5,7 +5,7 @@ import { bindActionCreators } from 'redux';
 import _ from 'lodash';
 
 import StatusMessage from '../components/StatusMessage';
-import AttorneyTable from './AttorneyTable';
+import JudgeReviewTaskTable from './JudgeReviewTaskTable';
 import AppSegment from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/AppSegment';
 import Alert from '../components/Alert';
 
@@ -13,52 +13,40 @@ import {
   resetErrorMessages,
   resetSuccessMessages,
   resetSaveState,
-  showErrorMessage
 } from './uiReducer/uiActions';
 import { clearCaseSelectSearch } from '../reader/CaseSelect/CaseSelectActions';
 
 import { fullWidth } from './constants';
 
-class AttorneyListView extends React.PureComponent {
+class JudgeReviewTaskListView extends React.PureComponent {
   componentWillUnmount = () => {
     this.props.resetSaveState();
     this.props.resetSuccessMessages();
-    this.props.resetErrorMessages();
   }
 
   componentDidMount = () => {
     this.props.clearCaseSelectSearch();
     this.props.resetErrorMessages();
-
-    if (_.some(this.props.tasks, (task) => !task.attributes.task_id)) {
-      this.props.showErrorMessage({
-        title: 'Some cases need preliminary DAS assignments',
-        detail: `Cases marked with exclamation points need to be assigned 
-        to you through DAS. Please contact your judge or senior counsel to 
-        create preliminary DAS assignments.`
-      });
-    }
   };
 
   render = () => {
-    const { messages } = this.props;
-    const noTasks = !_.size(this.props.tasks) && !_.size(this.props.appeals);
+    const reviewableTasks = {};
+    let reviewableCount = 0;
+    for (const k in this.props.tasks) {
+      if (this.props.tasks[k].attributes.task_type === "Review") {
+        reviewableCount++;
+      }
+    }
     let tableContent;
 
-    if (noTasks) {
+    if (reviewableCount === 0) {
       tableContent = <StatusMessage title="Tasks not found">
-        Sorry! We couldn't find any tasks for you.
+        Congratulations! You don't have any decisions to sign.
       </StatusMessage>;
     } else {
       tableContent = <div>
-        <h1 {...fullWidth}>Your Queue</h1>
-        {messages.error && <Alert type="error" title={messages.error.title}>
-          {messages.error.detail}
-        </Alert>}
-        {messages.success && <Alert type="success" title={messages.success}>
-          If you made a mistake please email your judge to resolve the issue.
-        </Alert>}
-        <AttorneyTable />
+        <h1 {...fullWidth}>Review {reviewableCount} Cases</h1>
+        <JudgeReviewTaskTable />
       </div>;
     }
 
@@ -68,16 +56,13 @@ class AttorneyListView extends React.PureComponent {
   };
 }
 
-AttorneyListView.propTypes = {
+JudgeReviewTaskListView.propTypes = {
   tasks: PropTypes.object.isRequired,
   appeals: PropTypes.object.isRequired
 };
 
 const mapStateToProps = (state) => ({
   ..._.pick(state.queue.loadedQueue, 'tasks', 'appeals'),
-  ..._.pick(state.ui, 'messages'),
-  ..._.pick(state.queue.pendingChanges, 'taskDecision'),
-  judges: state.queue.judges
 });
 
 const mapDispatchToProps = (dispatch) => ({
@@ -86,8 +71,7 @@ const mapDispatchToProps = (dispatch) => ({
     resetErrorMessages,
     resetSuccessMessages,
     resetSaveState,
-    showErrorMessage
   }, dispatch)
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(AttorneyListView);
+export default connect(mapStateToProps, mapDispatchToProps)(JudgeReviewTaskListView);
